LoessCyc Normalization

This function is a wrapper around limma::normalizeCyclicLoess() with method = "pairs". Each pair of columns is normalized mutually to each other. See this paper for more information. Also see limma::normalizeCyclicLoess().

Usage

loesscyc_normalize(exp, by = NULL, ...)

Arguments

exp

An experiment object.

by

A column name in sample_info to stratify by. Optional. If provided, the normalization will be performed within each group.

...

Additional arguments to pass to limma::normalizeCyclicLoess().

Value

An experiment object with the expression matrix normalized.